can now change floppy disk via the UI. more places would now deal with full 32bit address space. implemented the remaining IDE commands, at least a bit. Escape is now not the key for detaching the mouse.

+++ RunAmiga.Core/Memory/KickstartROM.cs
@@ -50,7 +50,7 @@ namespace RunAmiga.Core.Memory
 				{
 					throw new ArgumentOutOfRangeException();
 				}
-				addressMask = memoryRange.Length - 1;
+				addressMask = (uint)(memoryRange.Length - 1);
 			}
 			catch
 			{
